Fix remaining BX_UNBOXING findings in appendStructuredInstructionBody
SpotBugs' BX detector reports only one instance per method, so fixing
the first unbox/rebox at line 3827 surfaced three more in the same
method (the jump, switch-case and switch-default target lookups). All
cast an Integer to (int) only to rebox it for a Map<Integer,Integer>
get; drop the casts.

The two remaining (int) casts in the method are legitimate: one boxes a
freshly computed int as a Map key, the other is a List<Long> index.

+++ vm/ByteCodeTranslator/src/com/codename1/tools/translator/JavascriptMethodGenerator.java
@@ -4016,7 +4016,7 @@ final class JavascriptMethodGenerator {
                 if (instruction instanceof Jump) {
                     Jump jump = (Jump) instruction;
                     Integer t = labelToIndex.get(jump.getLabel());
-                    int tb = startToBlock.get((int) t);
+                    int tb = startToBlock.get(t);
                     String cond = structuredJumpCondition(jump.getOpcode(), ctx);
                     if (cond == null && jump.getOpcode() != Opcodes.GOTO) {
                         return _sb(method, instructions, "L3263");
@@ -4049,7 +4049,7 @@ final class JavascriptMethodGenerator {
                     Label[] swl = sw.getLabels();
                     for (int ki = 0; swl != null && ki < swl.length; ki++) {
                         Integer t = swl[ki] == null ? null : labelToIndex.get(swl[ki]);
-                        int tb = startToBlock.get((int) t);
+                        int tb = startToBlock.get(t);
                         if (entrySp[tb] < 0) {
                             entrySp[tb] = ctx.sp;
                         } else if (entrySp[tb] != ctx.sp) {
@@ -4060,7 +4060,7 @@ final class JavascriptMethodGenerator {
                                .append(structuredGoto(tb, b, loopEnd, tryStarts)).append(";\n");
                     }
                     Integer dT = sw.getDefaultLabel() == null ? null : labelToIndex.get(sw.getDefaultLabel());
-                    int db = startToBlock.get((int) dT);
+                    int db = startToBlock.get(dT);
                     if (entrySp[db] < 0) {
                         entrySp[db] = ctx.sp;
                     } else if (entrySp[db] != ctx.sp) {
